+++ scripts/build-website.bash
@@ -2,7 +2,6 @@
 set -e
 cd apps/website/
 bundle
-npm run build -- --incremental
+bundle exec jekyll build -d ../../origami.ft.com --incremental
 cd ../storybook/
-npm install
 npm run build-storybook