+++ scripts/component/test.bash
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+origami-build-tools test
+a11y
+validate-component-demo-html
+test-sass