+++ components/o-layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### [5.0.7](https://www.github.com/Financial-Times/origami/compare/o-layout-v5.0.6...o-layout-v5.0.7) (2021-11-02)
+
+
+### Bug Fixes
+
+* Update `o-brand` in components, replace "master" with "core" ([322617e](https://www.github.com/Financial-Times/origami/commit/322617ea80f30a6825d9c36872e05574b871ea82))
+
 ### [5.0.6](https://www.github.com/Financial-Times/origami/compare/o-layout-v5.0.5...o-layout-v5.0.6) (2021-09-24)